requirements.txt
google-adk>=1.12.0
yfinance>=0.2.0
streamlit>=1.28.0
python-dotenv>=1.0.0
pydantic>=2.0.0
google-genai>=0.2.0
requests>=2.31.0

